07_inference/requirements.txt
# Python dependencies for inference
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
rouge-score>=0.1.2
scikit-learn>=1.3.0
tqdm>=4.65.0
